+++ app/models/__init__.py
@@ -1,3 +1,4 @@
 from .db import db
-from .user import User, Coder, Review, Skill, Project, project_skills,coder_skills
+from .user import User
+from .new_model import Coder, Review, Skill, Project, project_skills, coder_skills
 from .db import environment, SCHEMA
